Little changes on optfrme heurisics and ready for CORS final experiments

+++ MyProjects/MODM/ProblemInstance.h
@@ -131,7 +131,7 @@ public:
 	{
 		return c;
 	}
-	int getHurdleRate()
+	double getHurdleRate()
 	{
 		return R;
 	}
